refactor: v2 UI and cleanup
