+++ components/easter-egg/button.vue
@@ -11,12 +11,7 @@ const CAT_ICON: Ref<typeof ICONS.PAW | typeof ICONS.CAT> = ref(ICONS.PAW);
 </script>
 
 <template>
-    <button
-        class="fixed bottom-0 left-0 mb-2 ml-2 opacity-20 hover:opacity-60"
-        @mouseover="CAT_ICON = ICONS.CAT"
-        @mouseout="CAT_ICON = ICONS.PAW"
-        @click="playNya"
-    >
+    <button @mouseover="CAT_ICON = ICONS.CAT" @mouseout="CAT_ICON = ICONS.PAW" @click="playNya">
         <Icon :name="CAT_ICON" size="2em" />
     </button>
 </template>
