+++ components/ui/Navigation.vue
@@ -1,15 +1,16 @@
 <script setup lang="ts">
+const { session } = useAuth();
 const router = useRouter();
-const pages = [
+const pages = ref([
     {
         label: "Map",
         icon: "heroicons:map-solid",
-        to: "/v2",
+        to: "/",
     },
     {
         label: "Gallery",
         icon: "heroicons:photo-solid",
-        to: "/v2/gallery",
+        to: "/gallery",
     },
     {
         label: "Awards",
@@ -19,9 +20,20 @@ const pages = [
     {
         label: "Credits",
         icon: "mdi:handshake",
-        to: "/v2/credits",
+        to: "/credits",
     },
-];
+    {
+        label: "Dashboard",
+        icon: "mdi:view-dashboard",
+        to: "/dashboard",
+        devs: true,
+    },
+]);
+
+const filterPages = (pgs: any[]) => {
+    if (session.value && ["256048990750113793", "769556133215862784"].indexOf(session.value.profile.id) !== -1) return pgs;
+    else return pgs.filter((value) => !value.devs);
+}
 </script>
 
 <template>
@@ -34,7 +46,7 @@ const pages = [
         </div>
 
         <div class="flex flex-col gap-2">
-            <NuxtLink v-for="page in pages" :to="page.to" active-class="text-primary-500">
+            <NuxtLink v-for="page in filterPages(pages)" :key="page.label" :to="page.to" active-class="text-primary-500">
                 <Icon :name="page.icon" size="1.5em" class="mr-2" />
                 <span class="text-lg font-medium">{{ page.label }}</span>
             </NuxtLink>
@@ -43,7 +55,10 @@ const pages = [
 
         <div class="flex flex-col items-start gap-2">
             <div class="flex flex-row items-center gap-2">
-                <EasterEggV2Button />
+                <EasterEggButton />
+                <a href="https://github.com/Tougrel/Pestoverse" target="_blank" class="hover:text-primary-700 dark:hover:text-primary-500 transition-colors">
+                    <Icon :name="ICONS.GITHUB" size="2em" />
+                </a>
                 <UiColorPicker size="2em" />
             </div>
             <p class="text-sm text-gray-400">
