+++ pages/gallery.vue
@@ -9,6 +9,7 @@ const images = gallery.flatMap((entry) =>
     })),
 ) as [{ name: string; src: string; width: number; height: number }];
 
+const loading = ref(true);
 const modalOpen = useState<boolean>("gallery-modal", () => false);
 const modalImage = useState<string>("gallery-modal-image");
 
@@ -16,11 +17,22 @@ const openModal = (image: string) => {
     modalOpen.value = true;
     modalImage.value = getFullImage(image);
 };
+
+definePageMeta({
+    description: "Photos of places the pestini have traveled around the world!",
+});
+
+onMounted(() => {
+    setTimeout(() => (loading.value = false), 1500);
+});
 </script>
 
 <template>
     <NuxtLayout name="default">
-        <div class="max-w-screen-2xl columns-1 gap-4 sm:columns-2 md:columns-3 lg:columns-4">
+        <div v-if="loading" class="flex h-full w-full flex-col items-center justify-center">
+            <img src="static/images/emotes/waddle.gif" decoding="async" loading="lazy" class="bg-cover bg-repeat-x" />
+        </div>
+        <div v-show="!loading" class="columns-1 gap-4 p-4 sm:columns-2 md:columns-3 lg:columns-4 xl:columns-5">
             <UiImage
                 v-for="image in images"
                 @click="openModal(image.src)"
